Moved the `TextShadow` component from `bevy_ui::ui_node` to `bevy_text::text`.
